chore(refactor): move git files under `git` module
