+++ crates/release_plz_core/src/git/backend.rs
@@ -1,5 +1,5 @@
 use crate::GitHub;
-use crate::{gitea_client::Gitea, gitlab_client::GitLab};
+use crate::git::{gitea_client::Gitea, gitlab_client::GitLab};
 
 use crate::pr::Pr;
 use anyhow::Context;
