+++ crates/release_plz_core/src/git/mod.rs
@@ -1,0 +1,4 @@
+pub mod backend;
+pub mod gitea_client;
+pub mod github_client;
+pub mod gitlab_client;
