+++ crates/release_plz_core/src/lib.rs
@@ -1,13 +1,10 @@
-mod backend;
 mod cargo;
 mod changelog;
 mod changelog_parser;
 mod clone;
 mod diff;
 mod download;
-mod gitea_client;
-mod github_client;
-mod gitlab_client;
+mod git;
 mod next_ver;
 mod package_compare;
 mod package_path;
@@ -22,12 +19,12 @@ mod tmp_repo;
 mod update;
 mod version;
 
-pub use backend::GitBackend;
+pub use git::backend::GitBackend;
 pub use changelog::*;
 pub use download::read_package;
-pub use gitea_client::Gitea;
-pub use github_client::GitHub;
-pub use gitlab_client::GitLab;
+pub use git::gitea_client::Gitea;
+pub use git::github_client::GitHub;
+pub use git::gitlab_client::GitLab;
 pub use next_ver::*;
 pub use package_compare::*;
 pub use package_path::*;
