+++ crates/release_plz_core/src/release_pr.rs
@@ -5,7 +5,7 @@ use git_cmd::Repo;
 use anyhow::{anyhow, Context};
 use tracing::{info, instrument};
 
-use crate::backend::{contributors_from_commits, GitClient, GitPr, PrEdit};
+use crate::git::backend::{contributors_from_commits, GitClient, GitPr, PrEdit};
 use crate::pr::{Pr, BRANCH_PREFIX};
 use crate::{
     copy_to_temp_dir, publishable_packages, update, GitBackend, PackagesUpdate, UpdateRequest,
